+++ webview_flutter/step_05/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:webview_flutter/webview_flutter.dart';
+
+import 'src/web_view_stack.dart';
 
 void main() {
   runApp(
@@ -24,47 +25,6 @@ class _WebViewAppState extends State<WebViewApp> {
         title: const Text('Flutter WebView'),
       ),
       body: const WebViewStack(),
-    );
-  }
-}
-
-class WebViewStack extends StatefulWidget {
-  const WebViewStack({Key? key}) : super(key: key);
-
-  @override
-  State<WebViewStack> createState() => _WebViewStackState();
-}
-
-class _WebViewStackState extends State<WebViewStack> {
-  var loadingPercentage = 100;
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        WebView(
-          initialUrl: 'https://flutter.dev',
-          onPageStarted: (url) {
-            setState(() {
-              loadingPercentage = 0;
-            });
-          },
-          onProgress: (progress) {
-            setState(() {
-              loadingPercentage = progress;
-            });
-          },
-          onPageFinished: (url) {
-            setState(() {
-              loadingPercentage = 100;
-            });
-          },
-        ),
-        if (loadingPercentage < 100)
-          LinearProgressIndicator(
-            value: loadingPercentage / 100.0,
-          ),
-      ],
     );
   }
 }
