+++ webview_flutter/step_08/lib/src/menu.dart
@@ -1,0 +1,38 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+
+enum _MenuOptions {
+  navigationDelegate,
+  // We will add more menu options here later
+}
+
+class Menu extends StatelessWidget {
+  const Menu({required this.controller, Key? key}) : super(key: key);
+
+  final Completer<WebViewController> controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<WebViewController>(
+      future: controller.future,
+      builder: (context, controller) {
+        return PopupMenuButton<_MenuOptions>(
+          onSelected: (value) async {
+            switch (value) {
+              case _MenuOptions.navigationDelegate:
+                controller.data!.loadUrl('https://youtube.com');
+                break;
+            }
+          },
+          itemBuilder: (context) => [
+            const PopupMenuItem<_MenuOptions>(
+              value: _MenuOptions.navigationDelegate,
+              child: Text('Navigate to YouTube'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
