+++ webview_flutter/step_08/lib/src/web_view_stack.dart
@@ -1,0 +1,63 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+
+class WebViewStack extends StatefulWidget {
+  const WebViewStack({required this.controller, Key? key}) : super(key: key);
+
+  final Completer<WebViewController> controller;
+
+  @override
+  State<WebViewStack> createState() => _WebViewStackState();
+}
+
+class _WebViewStackState extends State<WebViewStack> {
+  var loadingPercentage = 100;
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        WebView(
+          initialUrl: 'https://flutter.dev',
+          onWebViewCreated: (webViewController) {
+            widget.controller.complete(webViewController);
+          },
+          onPageStarted: (url) {
+            setState(() {
+              loadingPercentage = 0;
+            });
+          },
+          onProgress: (progress) {
+            setState(() {
+              loadingPercentage = progress;
+            });
+          },
+          onPageFinished: (url) {
+            setState(() {
+              loadingPercentage = 100;
+            });
+          },
+          navigationDelegate: (navigation) {
+            final host = Uri.parse(navigation.url).host;
+            if (host.contains('youtube.com')) {
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(
+                  content: Text(
+                    'Blocking navigation to $host',
+                  ),
+                ),
+              );
+              return NavigationDecision.prevent;
+            }
+            return NavigationDecision.navigate;
+          },
+        ),
+        if (loadingPercentage < 100)
+          LinearProgressIndicator(
+            value: loadingPercentage / 100.0,
+          ),
+      ],
+    );
+  }
+}
