+++ webview_flutter/step_10/lib/main.dart
@@ -1,6 +1,11 @@
 import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:webview_flutter/webview_flutter.dart';
+
+import 'src/menu.dart';
+import 'src/navigation_controls.dart';
+import 'src/web_view_stack.dart';
 
 void main() {
   runApp(
@@ -31,209 +36,6 @@ class _WebViewAppState extends State<WebViewApp> {
         ],
       ),
       body: WebViewStack(controller: controller),
-    );
-  }
-}
-
-class WebViewStack extends StatefulWidget {
-  const WebViewStack({required this.controller, Key? key}) : super(key: key);
-
-  final Completer<WebViewController> controller;
-
-  @override
-  State<WebViewStack> createState() => _WebViewStackState();
-}
-
-class _WebViewStackState extends State<WebViewStack> {
-  var loadingPercentage = 100;
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        WebView(
-          initialUrl: 'https://flutter.dev',
-          onWebViewCreated: (webViewController) {
-            widget.controller.complete(webViewController);
-          },
-          onPageStarted: (url) {
-            setState(() {
-              loadingPercentage = 0;
-            });
-          },
-          onProgress: (progress) {
-            setState(() {
-              loadingPercentage = progress;
-            });
-          },
-          onPageFinished: (url) {
-            setState(() {
-              loadingPercentage = 100;
-            });
-          },
-          navigationDelegate: (navigation) {
-            final host = Uri.parse(navigation.url).host;
-            if (host.contains('youtube.com')) {
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(
-                  content: Text(
-                    'Blocking navigation to $host',
-                  ),
-                ),
-              );
-              return NavigationDecision.prevent;
-            }
-            return NavigationDecision.navigate;
-          },
-          javascriptMode: JavascriptMode.unrestricted,
-          javascriptChannels: _createJavascriptChannels(context),
-        ),
-        if (loadingPercentage < 100)
-          LinearProgressIndicator(
-            value: loadingPercentage / 100.0,
-          ),
-      ],
-    );
-  }
-
-  Set<JavascriptChannel> _createJavascriptChannels(BuildContext context) {
-    return {
-      JavascriptChannel(
-        name: 'SnackBar',
-        onMessageReceived: (message) {
-          ScaffoldMessenger.of(context)
-              .showSnackBar(SnackBar(content: Text(message.message)));
-        },
-      ),
-    };
-  }
-}
-
-class NavigationControls extends StatelessWidget {
-  const NavigationControls({required this.controller, Key? key})
-      : super(key: key);
-
-  final Completer<WebViewController> controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<WebViewController>(
-      future: controller.future,
-      builder: (context, snapshot) {
-        final WebViewController? controller = snapshot.data;
-        if (snapshot.connectionState != ConnectionState.done ||
-            controller == null) {
-          return Row(
-            children: const <Widget>[
-              Icon(Icons.arrow_back_ios),
-              Icon(Icons.arrow_forward_ios),
-              Icon(Icons.replay),
-            ],
-          );
-        }
-
-        return Row(
-          children: <Widget>[
-            IconButton(
-              icon: const Icon(Icons.arrow_back_ios),
-              onPressed: () async {
-                if (await controller.canGoBack()) {
-                  await controller.goBack();
-                } else {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(content: Text('No back history item')),
-                  );
-                  return;
-                }
-              },
-            ),
-            IconButton(
-              icon: const Icon(Icons.arrow_forward_ios),
-              onPressed: () async {
-                if (await controller.canGoForward()) {
-                  await controller.goForward();
-                } else {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(content: Text('No forward history item')),
-                  );
-                  return;
-                }
-              },
-            ),
-            IconButton(
-              icon: const Icon(Icons.replay),
-              onPressed: () {
-                controller.reload();
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-enum _MenuOptions {
-  navigationDelegate,
-  userAgent,
-  javascriptChannel,
-}
-
-class Menu extends StatelessWidget {
-  const Menu({required this.controller, Key? key}) : super(key: key);
-
-  final Completer<WebViewController> controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<WebViewController>(
-      future: controller.future,
-      builder: (context, controller) {
-        return PopupMenuButton<_MenuOptions>(
-          onSelected: (value) async {
-            switch (value) {
-              case _MenuOptions.navigationDelegate:
-                controller.data!.loadUrl('https://youtube.com');
-                break;
-              case _MenuOptions.userAgent:
-                final userAgent = await controller.data!
-                    .runJavascriptReturningResult('navigator.userAgent');
-                ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content: Text(userAgent),
-                ));
-                break;
-              case _MenuOptions.javascriptChannel:
-                await controller.data!.runJavascript('''
-var req = new XMLHttpRequest();
-req.open('GET', "https://api.ipify.org/?format=json");
-req.onload = function() {
-  if (req.status == 200) {
-    let response = JSON.parse(req.responseText);
-    SnackBar.postMessage("IP Address: " + response.ip);
-  } else {
-    SnackBar.postMessage("Error: " + req.status);
-  }
-}
-req.send();''');
-                break;
-            }
-          },
-          itemBuilder: (context) => [
-            const PopupMenuItem<_MenuOptions>(
-              value: _MenuOptions.navigationDelegate,
-              child: Text('Navigate to YouTube'),
-            ),
-            const PopupMenuItem<_MenuOptions>(
-              value: _MenuOptions.userAgent,
-              child: Text('Show user-agent'),
-            ),
-            const PopupMenuItem<_MenuOptions>(
-              value: _MenuOptions.javascriptChannel,
-              child: Text('Lookup IP Address'),
-            ),
-          ],
-        );
-      },
     );
   }
 }
